+++ driver/web/conversions_auth.go
@@ -227,7 +227,8 @@ func serviceRegFromDef(item *Def.ServiceReg) (*model.ServiceReg, error) {
 		return nil, err
 	}
 	return &model.ServiceReg{Registration: authservice.ServiceReg{ServiceID: item.ServiceId, ServiceAccountID: serviceAccountID, Host: item.Host, PubKey: pubKey},
-		Name: item.Name, Description: item.Description, InfoURL: defString(item.InfoUrl), LogoURL: defString(item.LogoUrl), Scopes: scopes, FirstParty: item.FirstParty}, nil
+		Name: item.Name, Description: item.Description, InfoURL: defString(item.InfoUrl), LogoURL: defString(item.LogoUrl), VersionPath: item.VersionPath,
+		Scopes: scopes, FirstParty: item.FirstParty}, nil
 }
 
 func serviceRegToDef(item *model.ServiceReg) *Def.ServiceReg {
@@ -243,7 +244,7 @@ func serviceRegToDef(item *model.ServiceReg) *Def.ServiceReg {
 	pubKey := pubKeyToDef(item.Registration.PubKey)
 	scopes := serviceScopeListToDef(item.Scopes)
 	return &Def.ServiceReg{ServiceId: item.Registration.ServiceID, ServiceAccountId: serviceAccountID, Host: item.Registration.Host,
-		PubKey: pubKey, Name: item.Name, Description: item.Description, InfoUrl: &item.InfoURL, LogoUrl: &item.LogoURL,
+		PubKey: pubKey, Name: item.Name, Description: item.Description, InfoUrl: &item.InfoURL, LogoUrl: &item.LogoURL, VersionPath: item.VersionPath,
 		Scopes: &scopes, FirstParty: item.FirstParty}
 }